+++ libs/react/src/fetchers/axiosFetcher.ts
@@ -1,7 +1,7 @@
 import axios from 'axios';
 import { set } from 'lodash';
-import { useCookies } from '@coldpbc/hooks';
 import { resolveAPIUrl } from './helper';
+import cookies from 'js-cookie';
 
 const baseURL = resolveAPIUrl();
 
@@ -11,17 +11,9 @@ const baseURL = resolveAPIUrl();
  */
 export const axiosFetcher = (params: Array<string>) => {
   try {
-    // eslint-disable-next-line react-hooks/rules-of-hooks
-    const { getCookieData } = useCookies();
-    const cookie = getCookieData();
-
-    const token = cookie ? cookie.accessToken : null;
-    const user = cookie ? cookie.user : null;
-
-    if (!token)
-      throw new Error(
-        `${user?.email} attempted to request data from ${params[0]} however a token was not found in the cookie.`,
-      );
+    const raw = cookies.get('coldpbc');
+    const coldpbc = raw ? JSON.parse(raw) : null;
+    const token = coldpbc ? coldpbc.accessToken : null;
 
     const config = {
       baseURL: baseURL,