+++ libs/react/src/hooks/useCookies.ts
@@ -4,9 +4,8 @@ export const useCookies = () => {
   const setCookieData = (
     user: any,
     token: string,
-    expires = new Date(1000 * 60 * 60 * 24),
+    expires = new Date(Date.now() + 1000 * 60 * 60 * 24 * 30),
   ) => {
-    // Store the user and token in a cookie
     cookie.set(
       'coldpbc',
       JSON.stringify({ accessToken: token, expires, user }),
@@ -18,27 +17,35 @@ export const useCookies = () => {
     );
   };
 
-  return { setCookieData, getAccessToken, getCookieUser, getCookieData };
-};
+  const getCookieData = () => {
+    const raw = cookie.get('coldpbc');
+    const data = raw ? JSON.parse(raw) : null;
 
-export const getCookieData = () => {
-  // Get the data from cookie
-  const raw = cookie.get('coldpbc');
-  const data = raw ? JSON.parse(raw) : null;
+    return data;
+  };
 
-  return data;
-};
+  const getCookieUser = () => {
+    const data = getCookieData();
 
-export const getCookieUser = () => {
-  // Get the token from cookie
-  const data = getCookieData();
+    return data ? data.user : null;
+  };
 
-  return data ? data.user : null;
-};
+  const getAccessToken = () => {
+    // Get the token from cookie
+    const data = getCookieData();
 
-export const getAccessToken = () => {
-  // Get the token from cookie
-  const data = getCookieData();
+    return data ? data.accessToken : null;
+  };
 
-  return data ? data.accessToken : null;
+  const clearCookieData = () => {
+    cookie.remove('coldpbc');
+  };
+
+  return {
+    setCookieData,
+    getAccessToken,
+    getCookieUser,
+    getCookieData,
+    clearCookieData,
+  };
 };